Implement isSynchronized() in LocalNonWorkspaceTypedElement
To solve bug 481682, we'll use a LocalNonWorkspaceTypedElement to show
a diff on a file not in any open Eclipse project. The super class
implements isSynchronized() only for accessible workspace resources via
the Eclipse resource tree. We need to handle it ourselves here,
otherwise any attempt to save from the compare editor will pop up a
dialog falsely claiming the file had been modified outside of the
compare editor.

The implementation is closely modeled after the one in the super class,
but using the file system's last-modified timestamp directly instead of
going through the Eclipse resource tree, which would always give -1L
for such inaccessible resources.

Bug: 481682
Change-Id: Idb73c049f28327efdd65bcb0fdd6465e9928e317
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>