Update patch set 1
Patch Set 1:

> While I see what this does code-wise, could you describe what this
 > changes UI-wise? What works now that didn't work before?

Make History/Repositories/Staging view "link" to selection.

From the git repositories view open a file in the editor which is not in the workspace. The 3 Git views above will now show the appropriate content if you alternatively activates the "external" editor or select a  project in the packages explorer from a different repository.

Without the change, activating the editor doesn't change the content to the right repository in the Git views.

Patch-set: 1