Update patch set 1
Patch Set 1:

> > Git staging view doesn't update if I edit a file from a closed
 > > project, make some change, and save. Not sure if that was
 > supposed
 > > to be fixed by this.
 > 
 > Git staging view also doesn't show the unstaged change if after the
 > modification, I close the editor, select a different repo X in the
 > git repositories view, then again open the editor (from the git
 > repositories view) on that previously changed file. The staging
 > view does update: the staged/unstaged changes from repo X vanish,
 > but but the new unstaged change in the file just opened doesn't
 > appear in the staging view. So it looks as if that problem is due
 > again to there no IResourceChangeEvent being triggered, and thus
 > the IndexDiffCache isn't updated.

Thomas, this patch is not about updating any view state if a file modification state changes, it is about to *link* the view to the file/repository if the view is in "linked" mode, and to enable "Show In -> <view>" to work with the external file input if the view is not in "linked" mode.

I've tested and it works fine with Repository, History and Staging view (later of course only if it knows that file is changed).

Patch-set: 1
Reviewer: Gerrit User 1179 <1179@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0