Update patch set 1
Patch Set 1:

> Thomas, this is what you were looking before I guess :-)

More or less. Interesting approach. Works if the staging view is open. Doesn't work if the staging view is not open, file in closed project is edited (via git repositories view), and then saved. Then open staging view: unstaged change is still not shown. (No surprise.)

I think putting this in the staging view is not quite right. The fundamental problem is not that the staging view misses updates, that's just the symptom. The real problem is that the index diff doesn't get updated. So either this listener needs to be added more globally (activator?), or we find a more direct way to make editors on save also update the index if needed, even for files not in open projects.

Patch-set: 1