Update patch set 2
Patch Set 2:

(3 comments)

Indeed it was late yesterday.

I'm sorry, but I now think this is still not the right place. This is an additional listener to update the right IndexDiffCacheEntry. I actually think it belongs there, like the listener or resource change events. Yes, that's over in core, but note that none of the new imports are UI-related. So I think this can be moved safely over to org.eclipse.egit.core.internal.indexdiff.IndexDiffCacheEntry. Core would need a new plugin dependency on org.eclipse.core.filebuffers.

Patch-set: 2
Label: Verified=0