Update patch set 2
Patch Set 2:

> looks like this large change doing multiple changes could be split
 > into smaller changes which do one change each and build on top of
 > each other.

I didn't do so in this case because pulling out the new IResourceState alone gives a change of about +565/-358 with no clear indication of why it was done, and bullets 2 to 4 in above list are really small things (the numbers are misleading; xml is is rather verbose, which bloats the number of changed lines).

But ok, I'll take it apart.

Patch-set: 2