Update patch set 4
Patch Set 4:

> This menu entries are now context sensitive. So if there is nothing
 > to do for them they aren't shown. Do you mean they should be shown
 > but as "disabled"? I will check the code again tomorrow...

This doesn't work, I tried with a dirty file, the "Add to Index" is neither shown on team menu in package explorer nor in team menu in the Java editor opened for the modified class. Same for "Remove from Index" on a file with staged modifications and for a new class with "Ignore" action.

If this works reliably I am fine to only show the commands which make sense on the selected resource, though this may be tricky or expensive to compute on packages, folders and projects, we should avoid a performance penalty here

Patch-set: 4