Update patch set 4
Patch Set 4:

I had tested with a general project, and that worked. But indeed, this doesn't work with a Java project, or with any other contributed items that don't implement IResource (the type that ResourcePropertyTester tester is registered to).

Looks like I'll have to move the new property tests either into a separate tester registered on java.util.Collection (iterating inside the tester), or possibly move it to the SelectionPropertyTester.

That change has to happen in the parent commit of this one. The real problem is not this change here, but that the parent commit is wrong.  Looks like property testers don't play well with the adapter pattern: the core support for finding applicable testers only looks at the type hierarchy of the selected elements, not at their adaptations.

Back to the drawing board...

Patch-set: 4
Label: Verified=0