Synchronize View: show 'Add To Index' only when applicable
Also add 'Remove From Index', and make the menu entry order the same
for IResources and for GitModelObjects. Change the ResourceStateFactory
to work not only for IResource but also for java.io.File.

To handle unstaged deletions, propagate isMissing() through
IResourceState. (This fixes the _display_ of menu entries, but 'Add to
Index' doesn't work on unstaged deletions -- that's bug 483404.)

Bug: 355809
Change-Id: I2b111dd7be58c0b73d200278aa7f0fe2aaaccda9
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>