Try to make that EasymportWizardTest more stable
That test didn't run at all on my local machine; it failed already
because getting the branch info takes a bit more time from here.

At the end it failed because it tried to get the projects before the
import was done.

And on EGit Hudson, it usually fails because it cannot find
"File->Import...".

Add some more waiting all around.

Change-Id: I8f5a9cf61ffa55cfe12688b668563323ce4ef6bd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>