Try to make that EasymportWizardTest more stable
That test didn't run at all on my local machine; it failed already
because getting the branch info takes a bit more time from here.

Then it failed because it tried to get the projects before the
import was done.

Finally I had an exception reported by the Java auto-builder who
couldn't find the projects that were imported and then deleted at the
end of the test.

And on EGit Hudson, it usually fails because it cannot find
"File->Import...".

Add some more waiting, switch off auto-build, close the welcome page,
and try to ensure the Eclipse workbench window is activated (hopefully
that will solve the "File->Import..." not found problem; I couldn't
reproduce that locally.)

Change-Id: I8f5a9cf61ffa55cfe12688b668563323ce4ef6bd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>