Update patch set 1
Patch Set 1:

BTW if you haven't noticed yet: if we merge changes, the egit "master" build job is triggered with some delay, and if it runs in parallel with egit.gerrit job we will most likely have unexpected sporadic test failures in either of both. 

Therefore I try to not schedule new gerrit builds for ~1 hour after merges to master.

Just to make sure are not observing this issue.

Patch-set: 1