Update patch set 2
Patch Set 2:

This should solve the final outstanding case from bug 355809. Works in my tests, except for interactions with other bugs relating to this Synchronize View, for instance bug 483404. In my tests, this works otherwise well in the workspace model. In the "Git Commits" model, the display of the actions works, but on nodes for deleted files, things go a bit haywire: the 'Ignore' action, for instance, is disabled.

One part of the problem is GitAdapterFactory, lines 110ff, which constructs a bogus resource containing the absolute path in git (i.e., in my case /Users/<username>/Git/<repo>/<project>/<deleted_file>) as a _workspace local_ path, which then later on leads to strange effects because Eclipse takes the first component of such a local path as the project name and concludes that there is no project "Users". If I do create a project "Users" and share it with the same repo, then the 'Ignore' action is enabled, but doesn't do anything.

This is very broken, but hasn't got anything to do with this change, which is about the display of these actions. I intend to use ResourceState computed from File in the repositories view next, where the 'Add to Index' action also is generally enabled in the work tree.

Patch-set: 2
Label: Code-Review=0