Properly clean up caches when a configured repository is removed
When a repository is removed, it was never removed from the
IndexDiffCache, which in fact has no operation to remove anything. Only
the preference in ResourceUtil is updated. This of course in turn means
there's a hard reference in IndexDiffCache to the repo, and it thus also
remains in the RepositoryCache indefinitely (unless it is really
deleted, not just removed from the view).

Clean this up:

1. IndexDiffCache listens on changes to the preference storing the
   configured repositories and removes any stale cache entries upon
   changes.
2. Individual cache entries really clear all listeners, and cancel
   their jobs when disposed.
3. The RepositoryCache also listens on that preference and also removes
   stale entries proactively. Using weak references is a nice touch,
   but relying on the GC to clean up is not really necessary here.

Bug: 483664
Change-Id: I8c20d4e8821e6c5d9fa7e62191a79275bef58bce
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>