[performance] Remove ContainerTreeIterator and AdaptableFileTreeIterator
Don't hook any extra Eclipse based iterators to jgit tree traversal:

1) Eclipse IResource API can't handle "unusual" links (unresolved or
recursive), see bug 290318. This is a major functional deficiency for
repositories which make use of links.

2) The assumption that Eclipse IResource API traverses the file system
faster as java.io does not hold on Windows/Linux/Mac/NFS with modern
(7+) Java, see bug 484498 comments 10 and 14,15,17. The opposite is
true: especially on Linux the java.io access is considerably faster as
IResource API or what the ContainerTreeIterator makes of it.

AdaptableFileTreeIterator/ContainerTreeIterator use is replaced with
default FileTreeIterator, tests adopted or removed.

Bug: 484498
Change-Id: I026abee1706e6fbcd470cc7262b9fc1e59a7fe8b
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>