Update patch set 1
Patch Set 1:

(3 comments)

Does this work on Windows? Also, this only works for links outside the workspace. It doesn't handle links inside the workspace (in an open project): shows left (workspace) the contents of the linked-to file, right (index) the content of the link itself, i.e., the file name of the file being linked to.

Patch-set: 1