Update patch set 2
Patch Set 2:

We could make it configurable and by default do not allow editing symlinks. Also we could restrict changes to use a single line only for the link. I personally think that this is OK to edit links (we do it on command line, so why not in the IDE). Actually this was what I wanted to do initially and why I discovered this bug.

Patch-set: 2
Label: Verified=0