Update patch set 3
Patch Set 3:

> Seems to work on MacOS for valid links. Apparently also for broken
 > links.
 > 
 > However, I'm extremely wary of having these links editable in the
 > compare editor. I get very strange things and follow-up errors if I
 > edit a link in the compare editor and make it invalid in doing so.

See below, we had multiple ussues in j|egit.

 > And I didn't even try to do really bad things, like setting the
 > link content to multiple lines of text.

I've updated the patch to use only the first line of the text, works like a charm :-)
 
 > In particular:
 > [...]
 > 2. Link is shown in staging view as deleted. 

Bug 484498, I've rebased this patch on top of the patch for bug 484498.

 > Other case:
 > [...]
 > 4. In Eclipse, do a Reset hard on the parent commit (with the valid
 > link). Fails with
 > java.io.IOException: Could not delete file /Users/thomas/Git/testproject_b/link.txt

Bug 484491, patch is already merged to jgit.

 > I realize that these errors have nothing to do with this change
 > here. Maybe they're even due to my being on a Mac and having a
 > somewhat oldish git version 1.7.7.5 (Apple Git-26) there.

No, that was all our bugs :-)

 > Nevertheless I would advise against making link contents editable.
 > I don't think that's a good idea. Showing the link content is, but
 > not editing it. Let's make the compare editor read-only for links
 > in workspace or index.

I've grocked about it. I do not think this is bad: same way we could disallow users renaming or deleting regular files or directories or both, because sometimes they don't really mean it.

Thanks to git model in general, and with the fixes of the bugs above, this is pretty safe to "play" with links now.

Patch-set: 3