Update patch set 3
Patch Set 3:

> I'll try to track it down.

It's related to the encoding problem on Macs mentioned earlier. I get

  java.io.EOFException: Short read of block.
	at org.eclipse.jgit.util.IO.readFully(IO.java:249)
	at org.eclipse.jgit.treewalk.WorkingTreeIterator.readContentAsNormalizedString(WorkingTreeIterator.java:1047)
	at org.eclipse.jgit.treewalk.WorkingTreeIterator.contentCheck(WorkingTreeIterator.java:985)
	at org.eclipse.jgit.treewalk.WorkingTreeIterator.isModified(WorkingTreeIterator.java:909)
	at org.eclipse.jgit.treewalk.filter.IndexDiffFilter.include(IndexDiffFilter.java:227)
	at org.eclipse.jgit.treewalk.filter.AndTreeFilter$List.include(AndTreeFilter.java:163)
	at org.eclipse.jgit.treewalk.TreeWalk.next(TreeWalk.java:654)
	at org.eclipse.jgit.lib.IndexDiff.diff(IndexDiff.java:437)
	at org.eclipse.egit.core.internal.indexdiff.IndexDiffCacheEntry.calcIndexDiffDataIncremental(IndexDiffCacheEntry.java:529)
	at org.eclipse.egit.core.internal.indexdiff.IndexDiffCacheEntry.access$11(IndexDiffCacheEntry.java:509)
	at org.eclipse.egit.core.internal.indexdiff.IndexDiffCacheEntry$5.updateIndexDiff(IndexDiffCacheEntry.java:455)
	at org.eclipse.egit.core.internal.indexdiff.IndexDiffUpdateJob.run(IndexDiffUpdateJob.java:75)

triggered via LocalNonWorkspaceTypedElement.refreshRepositoryState(). The point is: I tried the above with my link to "äéü.txt". Changed it to point to "äéü.txt0". As a result, 14 bytes are written (decomposed form), not 11. WorkingTreeIterator then tries read those 14 bytes, but gets only 11 from FileTreeIterator.openInputStream because that normalizes back to NFC. And actually we already get back a fully composed string from Files.readSymbolicLink()... looking through the source code of sun.nio.fs.MacOSXFileSystemProvider, sun.nio.fs.UnixFileSystemProvider, sun.nio.fs.UnixPath, and sun.nio.fs.MacOSXFileSystem, it looks as if on a Mac, createSymbolicLink() always writes NFD, but readSymbolicLink always returns NFC. Yuck.

Not sure what's the best way to deal with this, especially since doing "ln -s äéü.txt link.txt" on the command line produces a link with on-disk size of 10 bytes, so it evidently is in NFC, not NFD.

Patch-set: 3