Update patch set 3
Patch Set 3:

> > I'll try to track it down.
 > 
 > It's related to the encoding problem on Macs mentioned earlier. I
 > get
 > 
 > java.io.EOFException: Short read of block.

As you noticed, change I264c597 is a fix for this. However, even without any special characters there's still another problem: if a link is made invalid and then valid again, it doesn't re-appear in the package explorer. Needs a manual refresh. I think this is due to our using a LocalNonWorkspaceTypedElement for a workspace resource. So LocalNonWorkspaceTypedElement would need to also refresh that resource.

Patch-set: 3