Update patch set 1
Patch Set 1:

> I can't reproduce the problem this is supposed to fix. Even without
 > this change, the Team menu only displays "Apply Patch..." and
 > "Share Project..." on unshared projects. Could it be that this was
 > already resolved by change I1b162615? Or maybe I misunderstood the
 > commit summary...

No, I don't think you misunderstood. I had this workaround on a local branch for a while, and it seems that the situation changed. The Gitflow UI plug-in now seems to be in ACTIVE state right after starting Eclipse is loaded (I started from within the IDE, and checked with the OSGi console). I'm not sure, though, if it is activated for the right reason, or if this is a bug.

I also tested this with the parent of I1b162615, and observed no difference, so that doesn't seem to have changed it.
I'll investigate this a bit more with the nightly builds, and get back to you.

Patch-set: 1
Reviewer: Gerrit User 1638 <1638@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0