Update patch set 2
Patch Set 2:

> with this change "start release" is enabled in history view if
 > gitflow wasn't initialized, this seems wrong

It seems to me that was introduced to master previously to this commit.
To be precise, I mean the visibility of the Gitflow sub-menu plus the "start release from ...".
With this commit, the Gitflow sub-menu is still there, but the "start release from ..." is greyed out for non-initialized repositories.

I also think it is wrong to show the sub-menu, but I would fix it separately, if you don't mind.

Patch-set: 2
Label: Verified=0