Avoid expensive property tester evaluations
The GitFlow property tester tests the presence of certain values in the
git config. Accessing the git config incurs three file accesses to get
the last-modified times of the system, user, and repo config files to
determine whether they have changed, and if so, to reload them. These
are potentially slow operations involving file I/O. Even just getting
the last-modified time apparently can be expensive, for instance if the
user config is on a non-local file system.

The GitFlow property testers were based on the current selection and
thus were re-evaluated by the Eclipse framework whenever the current
selection changed, including text selections. The text selection may
change on each keypress. This could lead to very noticeable delays if
access to the last-modified time of a config file was slow.

I don't see an easy way to improve the handling of config files. The
other approach to minimize the overhead of property testers is to
minimize the number of times they're re-evaluated. Since the GitFlow
property testers are not really interested in the current selection
but in a derived property of that (the current repository), this
change provides this current repository as a variable in the
IEvaluationContext and makes the property tester use that new variable.
As a result, the tests are only re-evaluated when the current selection
changes such that the current repository changes. Keypresses and text
selection changes in an editor no longer trigger re-evaluations of the
GitFlow property tests.

The current repository is still determined on each selection change.
However,

* this is normally a relatively fast operation as the repository is
  already cached most of the time, 
* the repository is determined only once in the source provider instead
  of again and again in each property test.

Bug: 484785
Change-Id: Id305fb54a6666f8e9715dcf0adb070880e1c098a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>