Update patch set 4
Patch Set 4:

So now the computation is in the background but the tester (and menus) will always see the right value if requested - they will need of course wait the first time, but at least it can cancel this.

I've disabled computation because of changed text selections now - this is an overkill. Selections in explorers etc will still work.

I'm still wondering if we should move this code to the gitflow plugin.

Patch-set: 4