Update patch set 9
Patch Set 9:

Sorry about the late reply; I was travelling.

I'm not sure this is the right approach.

1. I think something is fundamentally wrong if the property tester needs to know the source provider. (Yes, I see this has changed slightly in the follow-up change, but basically only the access to the RepositorySourceProvider has been hidden in a utility class.)
2. The text selection originally was included to have correct enablement of GitFlow submenu entries is the context menu in an open editor. To avoid that we re-evaluate the property tester on each text selection change, the RepositorySourceProvider was introduced. Now that we don't handle the text selection anymore, we wouldn't need any of this at all. We could just revert back to commit a612318d (before the RepositorySourceProvider was introduced) and simply remove the text selection handling in SelectionPropertyTestser.getRepository().
3. Not handling the text selection breaks the context menu in open editors. I don't get any GitFlow entries at all anymore. Yes, I realize that this is fixed in the follow-up change I0bc7aee.

The RepositorySourceProvider was supposed to do the cheap part (getting the repository from the selection), and enable us to do the expensive part (accessing the repo config) only if the repo really changed, which isn't the case as long as only the text selection within an editor changes. The assumption was that getting the repo from the selection was indeed cheap. We use that same mechanism elsewhere, and it never caused any performance problems. It's used in the lightweight decorator, and also in the ResourceStatePropertyTester and in the EGit UI SelectionPropertyTester (resourcesSingleRepository).

So, the real question is why is getting the repo from the selection in the RepositorySourceProvider suddenly expensive, when it isn't in other places?

Maybe it's because the RepositorySourceProvider gets instantiated rather early, and the projects are not mapped yet. That's what I read from Andrey's stack trace. If so, there might be ways to mitigate this:

* "Activating" the RepositorySourceProvider only once the rest of Eclipse startup is done. Let it return a null entry initially, and switch it on only once the workbench has been restored. At that point presumably the projects are mapped... I don't know off-hand how we'd detect "workbench has been restored", though.
* Alternatively, start a background job as soon as the EGit UI plugin is activated to asynchronusly try to map projects/restore the GitProviders, and hope that by the time the RepositorySourceProvider gets called, the projects are mapped. Perhaps it's even good enough to just try to load the project persistent properties in the background -- I didn't look into what gets ached there when and how.
* Alternatively, let the RepositorySourceProvider not handle text selections, but let it additionally listen to active part changes, and whenever an editor becomes active, start listening to changes to the editor input, then determine the repo from the editor input. Again, I don't know off-hand how we'd listen to either (active part changes and editor input changes), but I guess Eclipse has the mechanisms we'd need to do that. I'm also not sure if listening to editors is enough -- what about views?

Off-hand, all this job logic here strikes me as rather complicated -- even if in the end, with the follow-up change, it appears to work.

Patch-set: 9