Update patch set 9
Patch Set 9:

(2 comments)

This change mixes two things:

* Fixing the handling of text selections in the RepositorySourceProviderso that GitFlow submenu entries work correctly in the context menu of editors.
* Using the "current" repository from the RepositorySourceProvider in the rest of EGit.

The latter was something I had planned, too -- which was my reason for placing the RepositorySourceProvider in egit.ui and not in the gitflow.ui bundle ;-).

Can't we split that; or even better, include the text selection fix in I60aac46 so that that change becomes a complete solution?

Patch-set: 9