Wizard to configure pull operation
This is accessible on project's context menu, as "Pull..." just below
"Pull".

The wizard enables the user to specify the remote or add a new one, and
also to specify the remote reference. There is content assist for
published remote references, but the user may also pull from
non-published remote refs, such as pull requests on Github or patch sets
in Gerrit.

Bug: 485124
Change-Id: I88122d3fedf10de35a0c1c233b7dd9920bb8c0c7
Signed-off-by: Mickael Istria <mistria@redhat.com>
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>