Update patch set 4
Patch Set 4:

> 4 & 5. It seems to me that EGit always store the configuration of
 > 1st operation if none is already available for current branch
 > (independently of whether the operation succeeds or not). It would
 > most likely do the same if you use the "Push branch" wizard. If you
 > want to override it, you have to tick the checkbox. So as far as I
 > understand your comments, it's mostly that you'd like the checkbox
 > to be ticked and the remote configuration overwritten by default?

I thought I did that on a branch that was already set up to track origin/master, and I left that checkbox unchecked, so the attempt to pull from "foo" should not have overwritten the existing upstream config. I'll have to check more closely what's up.

And it makes no sense for EGit to store an upstream config that cannot be pulled from. So no, I don't want that checkbox be ticked by default. If EGit really unconditionally stores the first remote ref a fetch is attempted from as upstream even if that fetch fails, then that's maybe a separate bug. But again, I'll have to re-examine this case more closely to see what exactly happened.

Patch-set: 4