Update patch set 4
Patch Set 4:

> And it makes no sense for EGit to store an upstream config that cannot be pulled from. So no, I don't want that checkbox be ticked by default. If EGit really unconditionally stores the first remote ref a fetch is attempted from as upstream even if that fetch fails, then that's maybe a separate bug. But again, I'll have to re-examine this case more closely to see what exactly happened.

I disagree on that. It's not because it doesn't work (at that time) that there is no interest in storing it if user has decided so.
Let's assume users have the checkbox ticked, and simply don't have a connection at that time, or that you whole configuration fails for a typo; then I believe they expect it to be saved anyway and to be able to fix this later rather than having to rewrite everything.

Patch-set: 4