Update patch set 4
Patch Set 4:

Next patch set fixes 1&2.
About 3&6, I believe we need to think more about this. I'm not sure warning user of something possible is doable. Actually, the error shown in case of wrong reference is clear enough to present it to users IMO.
About 4&5, I gave a few tries and couldn't reproduce the issue you mentioned. I have the impression that the handling of the default upstream config is consistent.

Patch-set: 4