Update patch set 5
Patch Set 5:

To clarify: for me, having the warnings in the console output would also be OK. It's just that there are a lot of spurious ones, like "... new Git(...)" leading to a resource leak warning, or warnings due to null-annotations. Until JDT can properly deal with multiple null-annotations from different sources, we probably can't get a zero-warnings build.

Patch-set: 5