Minimize accessing the persistent property for the GitProvider
Accessing persistent properties may be expensive, and may block if
other threads manipulate the same persistent property bucket. (For
instance, JDT initialization does set persistent properties for source
attachments.)

Therefore, if we didn't find a GitProvider through the session property,
let's first check whether the project is already scheduled for mapping
in the MappingJob. If so, we do know:

* the project is indeed a Git project (some thread tested the persistent
  property already), and
* we don't need to test the persistent property again, and
* we don't need to try to schedule the project for mapping again.

This should significantly reduce the number of accesses to that
persistent property for projects that indeed have a GitProvider.

To reduce the number of persistent property accesses for unshared
projects or projects that are shared, but not with EGit, implement
logic similar to RepositoryProvider.markAsUnshared() to cache
information on the mapping in a session property. Note: this works
because mapping a repository provider creates the session property
first, and we reset our own property whenever we do find a provider
in that session property (any one will do).

Bug: 486081
Change-Id: I5272b0676921f7a6d59ca2445b5762e7b747a13c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>