Update patch set 2
Patch Set 2:

(2 comments)

@Matthias: in a follow-up change I plan to move the screenshots from their current default location at ${project.basedir}/screenshots to ${project.basedir}/target/screenshots. The current location has two problems:

1. It's is never cleared in a build, and thus we accumulate screenshots from very old builds. On Hudson, one doesn't see the last-modified date, so it's rather painful to find new screenshots.

2. Existing screenshots can get overwritten by subsequent builds.

Moving them to target/screenshots would solve (1), and then we could archive them as artifacts in the Hudson builds (so they don't get lost on the next build), which would also solve (2). If we do that, somebody would have to manually remove the old screenshots directories. (Or temporarily add a shell command to the initial build step to have them removed on the next run.)

I have added the new runner here as a test balloon -- want to see what Mickael thinks about it. If SWTBot should get updated like this, we might perhaps use a nightly build of SWTBot to have this. Or keep it here as is until there's an official SWTBot release that would include this.

Patch-set: 2
Label: Verified=0