Improve performance of SelectionUtils.getRepository()
Instead of converting selected elements to an IPath location always
and then using RepositoryMapping.getRepository(IPath), try to keep as
much info as needed and as possible: if the selection contains
IResources, keep them and use the more efficient
RepositoryMapping.getRepository(IResource).

In my tests on a test workspace with 46 open projects and the git
toolbar and global menu enabled, this results in a speedup of the
SelectionUtils.getRepository(IStructuredSelection) operation by a
factor of about 10 in many cases.

This could be optimized a little bit more if we could, instead of using
IResources, collect only their containing IProjects and then use
RepositoryMapping.getProject(), but I'm currently unsure whether that
would actually be correct. See the review comments at
https://git.eclipse.org/r/#/c/65510/ .

Bug: 486857
Change-Id: Ied35b304c8496c501dbf8109e1c8d5403b962efb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>