Improve performance of SelectionUtils.getRepository()
Instead of converting selected elements to an IPath location always
and then using RepositoryMapping.getRepository(IPath), try to keep as
much info as needed and as possible: if the selection contains
IResources, use their projects and the much more efficient
RepositoryMApping.getRepository(IProject).

In my tests on a test workspace with 46 open projects and the git
toolbar and global menu enabled, this results in a speedup of the
SelectionUtils.getRepository(IStructuredSelection) operation by a
factor of 10 to 20 in most cases.

Bug: 486857
Change-Id: Ied35b304c8496c501dbf8109e1c8d5403b962efb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>