Update patch set 4
Patch Set 4:

> Still, wondering if we should change the way how RepositoryMapping
 > getMapping(@NonNull IPath path) is implemented.

Agreed, but maybe not in this change.

 > So a map between Repository <-> Set<IProject> would be interesting,
 > why not in GitProjectData? BTW I can't get the design rationale
 > behind this class. The code seems to be "wild grown" over years
 > without clear idea behind.

So I'm not alone. This class has me puzzled from the very beginning. In particular GitProjectData.getRepositoryMapping(): this walks up the whole IContainer hierarchy. Can we ever have a case where a project is in one git repository, but a container inside the project is in some other git repository? Perhaps with linked resources...

Patch-set: 4