Update patch set 5
Patch Set 5:

> Which leaves me still a bit puzzled because RepositoryMapping.getMapping(IPath)
 > considers projects only. Digging around more in the early history
 > (this time in the current repo) shows:
 > 
 > * commit dfbdc45 (Shawn, 2009-09-30): the very initial version of
 > getMapping(IResource) did actually use the GitProjectData to obtain
 > the mapping, i.e., did do this container hierarchy walk.
 > * commit 610529c (Dariusz, 2011-11-18): then it was augmented to
 > also handle non-workspace resources, where the first version of the
 > current getMapping(IPath) was introduced. (Scanning all IProjects.)
 > * commit e65101e (Robin, 2012-08-26): refactoring introducing the
 > first version of the current getMapping(IPath). At that point
 > getMapping(IResource) still used the original code walking the
 > container hierarchy, and getMapping(IProject) didn't exist yet.
 > * commit a3da195 (Andrey, 2015-03-29): more refactoring,
 > getMapping(IProject) introduced, and getMapping(IResource) no
 > longer calling GitDataProvier.getRepositoryMapping(resource) but
 > getMapping(resource.getProject()).

The reason for this last change was: if neither the resource nor any of the parents are linked (the check was added in 199f7e), we only need to check the project for mapping. From 199f7e the code did nothing except wasting time. 

The original code expected to return a different repository for a submodule inside a project mapped to another one, but this seem to be broken (or never properly implemented) in GitProjectData and/or RepositoryMapping.

I think the hierarchy walk code could be removed after 199f7e. I've tried to step one commit back after 199f7e and also tried to partly revert a3da195 but I could not see any difference in behavior of submodules, or I do not know what to look for.

I think currently nobody really knows how GitProjectData/RepositoryMapping are supposed to work and what of their original functionality is still needed.

Both classes probably need to be cleaned up and their functionality/responsibility redesigned. 

 > Since that commit, we no longer do the container hierarchy walk.
 > 
 > Now I _do_ wonder:
 > * was this container hierarchy walk ever needed?

Before 199f7e (egit 2.3) - eventually, because it returned different repos for *linked* resources. Still I do not know how useful this was...

The main problem of all this linked/sub repositories is that being a "part" of another project they only confuse user / UI. I think original idea was to provide most close related context to the current file.

 > * if it is or was, what obscure use case no longer works since
 > a3da195, or if that didn't break anything, not even any obscure use
 > case, why did it not?

The only candidate would be something related to submodules but I don't see what it should be what was not broken (or simply never working) before already.

 > * the project scanning approach in getMapping(IPath) doesn't look
 > right to me; I would rather use the RepositoryCache to consider all
 > configured repositories, which would be feasible if it was made to
 > load on demand (I do have a change for that that I could
 > resurrect).

Probably the original author just had only few projects in his workspace. 100+ projects are not often to see in open source, except one imports entire eclipse git server content.

 > The approach in this change here (going right to the IProject from
 > an IResource) is only correct if that container hierarchy walk
 > isn't needed. So I'm currently unsure whether this change here is
 > good indeed. Otherwise, a mixture should be used: still
 > differentiate IResource and IPath, then use RepositoryMapping.getMapping(IResource)
 > or getMapping(IPath). This has two advantages:
 > 
 > 1. It delegates the whole question of whether this container
 > hierarchy walk is needed to the implementation of RepositoryMapping.
 > 2. Avoid duplicating the isLinked() handling I introduced
 > (RepositoryMApping alreadydoes that.
 > 
 > I'll push an amended change. Can you shed any light on the
 > questions above?

I tried.

Patch-set: 5