Update patch set 5
Patch Set 5:

> I tried.

Thanks :-)

So, in the current EGit 4.2. implementation we assume (except for linked resources) that all project content is in the same repo as the project itself, and PS 4 would thus be fine, too?

PS 4 is a bit more efficient (less distinct IProjects than distinct IResources in most cases), so if the duplicate isLinked() logic doesn't hurt, I'm also fine with reverting back to PS 4.

 > Both classes probably need to be cleaned up and their functionality/responsibility redesigned.

Looks like. Unfortunately it also looks like a Herculean task to figure out which bits still are needed (and what for exactly) and which not. I usually have neither linked resources nor submodules in my git-managed projects, so I might easily overlook something...

Patch-set: 5