Update patch set 5
Patch Set 5:

Some experimentation shows that assuming everything under a project was in the same repo as the project itself is just plain wrong in the presence of submodules. It causes (amongst other things) bug 446344, comment 11.

Moreover, the previous SelectionUtils.getRepository() implementation converting everything to IPath and then looking for the repo and in doing so scanning just the projects exacerbates the problems with submodules, because it fails to find a submodule repo if there's no project from the submodule in the workspace. (For instance, only a project from the parent repo imported.)

RepositoryMapping.getRepository() for IResource must not skip the container hierarchy walk, and the mappings for containers must be in GitProjectData if we want submodules to work.

So, this change here is more than just a performance improvement; it's also a prerequisite to fixing this submodule mess. (At least for IResources. Have no clear idea yet how to get this submodule stuff right for non-workspace files -- outside of any project, or in closed projects... one thing after another.)

Patch-set: 5
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0