Update patch set 1
Patch Set 1: Code-Review+1

Looks good and does enable the handler as wanted. But could we have a test for this, please? Should test that the handler is enabled when branch checked out as well as in detached HEAD state, and is disabled when there is no HEAD.

Patch-set: 1
Label: Code-Review=+1