Handle relative paths in RepositoryMapping creator
file.getAbsolutePath() with a relative path here resolves against
the current working directory of the Eclipse instance. The path is,
however, relative to the IContainer being mapped!

This causes many of the various occurrences of errors in the log about
gone repository mappings, especially those where the absolute path is
nowhere near any expected directory.

Bug: 456799
Bug: 476011
Change-Id: I9eebf4cb81b7584936f823885db8a5d9340035fd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>