Make AddToIndex/RemoveFromIndex work for multiple repositories again
Commit 4e89050 inadvertently introduced the restriction that all
resources in the selection be in one repository. However, AddToIndex,
RemoveFromIndex, and Ignore previously also worked if projects from
different repositories were selected.

Remove the restriction on a single repository from the
ResourceStatePropertyTester, and simplify the isEnabled() of AddToIndex
and RemoveFromIndex action handlers a bit.

The ResourceStatePropertyTester is currently used only for these
commands in the Team context menu. If is ever to be used with some
command for which all resources really must be from one repository, it
can be combined with the "resourcesSingleRepository" property test.

Performance should not suffer as all repository or index diff operations
in the property tester operate on caches.

Add new tests to test the "Add To Index" and "Remove From Index" actions
for single and multiple repository selections.

Change-Id: I71b1ece61cceeb08ee775e7afae33beee2b57b9f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>