WIP: Git 2.5 worktree: accept working-tree created with native git with "git worktree add"
JGit change "https://git.eclipse.org/r/#/c/54404/" adds the evaluation
of ".git" file with "gitdir" link to worktree and evaluates commondir.

With this chnange EGit should not check or expect the file "config" is
located in GIT_DIR, because in case of worktree it's located in
GIT_COMMON_DIR.

Generall question is why should EGit check the existence of some files
located in protected ".git" folder (database)? This task should be done
by JGit, e.g. validate the mapped repository.

I think this remove does not break any assumptions in EGit and if needed
the repository validation should be part of new task. 

Bug: 477475
Change-Id: Id1a54e0efb6fb7334ea4cb4a5507ed3e0bff0f7e
Signed-off-by: Andre Bossert <anb0s@anbos.de>