Test stability: SynchronizeViewGitChangeSetModelTest
The test was looking for a tree node with label "GeneralProject". As
the screenshot from unstable build 8131[1] clearly shows, the node is
labeled "> GeneralProject".

Looks like there's some asynchronous decoration going on, and depending
on whether the test tried to get the node before or after the decorator
ran it would succeed or fail.

Remedy by using TestUtil.getNode(), which only requires a substring
match, instead of SWTBotTreeItem.getNode(), which requires an exact
match. Also relax the assertion on the "<working tree>" node (same as
in the following test method).

There may be more similar problem spots lurking in here.

[1]
https://hudson.eclipse.org/egit/job/egit.gerrit/8131/artifact/org.eclipse.egit.ui.test/target/screenshots/shouldRefreshSyncResultAfterWorkspaceChange%28org.eclipse.egit.ui.view.synchronize.SynchronizeViewGitChangeSetModelTest%29.jpeg

Change-Id: If0051411196ac0148f2a6289d1dfba1e3d69f61d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>