[WIP] Test stability: GitSubscriberMergeContextTest
I can't get that test to fail locally; neither on native OS X not on a
Linux (Ubuntu 15.04) VM. Yet it fails frequently on Hudson. I don't see
anythign wrong with the test either. Let's include a lot of debug output
in an attempt to figure out what's going on.

Change-Id: Icdd054c5b7f878771046706f615746e3f47e9b4f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>