Update patch set 7
Patch Set 7:

(17 comments)

Hello Matthias,

Thank you for your review and sorry for the delay of reaction on my side, I've been unable to work on that earlier.
Here is a new patch set with fixes according to your comments. I tried to improve the ThreeWayDiffEntry.scan() method a bit, please tell me if it's ok for you.

Patch-set: 7
Label: Verified=0